write queryData function to eliminate code duplication
