+++ pages/models/model/[slug].tsx
@@ -2,8 +2,7 @@ import { ModelTypes, MODELS_QUERY_STRING } from "utils/queries/models";
 import { Layout } from "components/layout/PageLayout";
 import { NextPage } from "next";
 import { Model } from "components/models/Model";
-import { GRAPHQL_ENDPOINT } from "config";
-import { parseCookies } from "nookies";
+import { queryData } from "utils/functions/queryData";
 
 const title = `Models`;
 const description = `Whether it plane, car or tank, its all here!`;
@@ -22,34 +21,15 @@ const ModelPage: NextPage<iModelPage> = ({ model }) => {
 
 ModelPage.getInitialProps = async (ctx) => {
   const { slug } = ctx.query;
-  const cookies = parseCookies(ctx);
 
-  const options = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      query: MODELS_QUERY_STRING,
-      variables: { where: { slug: slug } },
-    }),
-  };
+  const data = await queryData(ctx, MODELS_QUERY_STRING, { slug: slug });
 
-  // inject the bearer token if its present
-  if (cookies?.token) {
-    // @ts-ignore
-    options.headers["Authorization"] = `Bearer ${cookies.token}`;
-  }
-
-  const response = await fetch(GRAPHQL_ENDPOINT, options);
-  const data = await response.json();
-
-  if (data?.data?.models?.length < 1) {
+  if (data?.models?.length < 1) {
     ctx?.res?.writeHead(301, { Location: "/404" });
     ctx?.res?.end();
     return { model: {} };
   } else {
-    return { model: data.data.models[0] };
+    return { model: data.models[0] };
   }
 };
